+++ driver-core/src/test/java/com/datastax/driver/core/PreparedStatementTest.java
@@ -20,6 +20,8 @@ import java.util.*;
 import static org.testng.Assert.assertFalse;
 
 import org.testng.annotations.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
@@ -426,5 +428,27 @@ public class PreparedStatementTest extends CCMBridge.PerClassSingleNodeCluster {
         BatchStatement batch = new BatchStatement();
         batch.add(ps.bind("k"));
         session.execute(batch);
+    }
+
+    @Test(groups="short")
+    public void should_set_routing_key_on_case_insensitive_keyspace_and_table() {
+        session.execute("CREATE TABLE ks.foo (i int PRIMARY KEY)");
+
+        PreparedStatement ps = session.prepare("INSERT INTO ks.foo (i) VALUES (?)");
+        BoundStatement bs = ps.bind(1);
+        assertThat(bs.getRoutingKey()).isNotNull();
+    }
+
+    @Test(groups="short")
+    public void should_set_routing_key_on_case_sensitive_keyspace_and_table() {
+        session.execute("CREATE KEYSPACE \"Test\" WITH replication = { "
+            + "  'class': 'SimpleStrategy',"
+            + "  'replication_factor': '1'"
+            + "}");
+        session.execute("CREATE TABLE \"Test\".\"Foo\" (i int PRIMARY KEY)");
+
+        PreparedStatement ps = session.prepare("INSERT INTO \"Test\".\"Foo\" (i) VALUES (?)");
+        BoundStatement bs = ps.bind(1);
+        assertThat(bs.getRoutingKey()).isNotNull();
     }
 }