+++ driver-core/CHANGELOG.rst
@@ -3,7 +3,7 @@ CHANGELOG
 
 2.0.8:
 ------
-- [bug] Fix token awareness for case-sensitive keyspaces (JAVA-526)
+- [bug] Fix token awareness for case-sensitive keyspaces and tables (JAVA-526)
 
 
 2.0.7: